Block forbidden globals like process in Node.js
One worry that remains, given that the code is executed in a separate
process that can be set with a zero-rights user,
is that the latter, setting to a restricted user, is not done.

The main source of risk in this case is the ability to use
the global process object, which includes methods such as dlopen()
and getgroups(), which can leak data and access dangerous resources.

We thus restrict access to a whitelisted set of global variables.
This gives defense in depth.

We perform this by crippling the globals from within the VM environment,
before the code is run.
Merely setting them as variables `var process;` would not be enough,
because access to the global object can be restored by the use of:

    var process = ({}).constructor.constructor("return process");

However, setting the global variable to null is an irreversible
loss of access:

    this.process = null;